float

+++ scripts/calcldavlb.py
@@ -99,7 +99,7 @@ if __name__ == '__main__':
     parser.add_argument("k_min", help="minimum number of topics for calculation of maximized variational lower bound", type=int)
     parser.add_argument("k_max", help="max number of topics for calculation of maximized variational lower bound", type=int)
     parser.add_argument("-s", "--kstep", help="step number of topics for calculation of maximized variational lower bound", type=int, default=1)
-    parser.add_argument("-d", "--conv_det", help="convergence determination for vb algorythm", type=double, default=0.001)
+    parser.add_argument("-d", "--conv_det", help="convergence determination for vb algorythm", type=float, default=0.001)
     args = parser.parse_args()
     calculator = VLBCalculator(args.BOW_filename, args.k_min, args.k_max, args.kstep, args.conv_det)
     calculator.run()
